refactor: 키워드 제거, 리뷰 관련 테스트 정리
