--- backend/src/main/java/reviewme/keyword/domain/Keyword.java
@@ -1,30 +1,0 @@
-package reviewme.keyword.domain;
-
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import jakarta.persistence.Table;
-import lombok.AccessLevel;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Entity
-@Table(name = "keyword")
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-@Getter
-public class Keyword {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @Column(name = "content", nullable = false)
-    private String content;
-
-    public Keyword(String content) {
-        this.content = content;
-    }
-
-}
